NF: refactor onCreate of AppCompatPreferenceActivity
Both class that inherited it had a lot of code in common
